# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: Search, filter, sort, and page through expenses
      parameters:
        - name: search
          in: query
          required: false
          description: Case-insensitive match on description, notes, category, or project name
          schema:
            type: string
            maxLength: 80
        - name: from
          in: query
          required: false
          description: Include expenses on or after this date
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        - name: to
          in: query
          required: false
          description: Include expenses on or before this date
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: categoryId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: labelId
          in: query
          required: false
          description: Only expenses carrying this label
          schema:
            type: string
            format: uuid
        - name: paymentMethod
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ExpensePaymentMethod"
        - name: currencyId
          in: query
          required: false
          description: Filter expenses by currency
          schema:
            type: string
            format: uuid
        - name: locationId
          in: query
          required: false
          description: Filter expenses by location
          schema:
            type: string
            format: uuid
        - name: parentCategoryId
          in: query
          required: false
          description: Filter expenses by top-level category, including all of its subcategories
          schema:
            type: string
            format: uuid
        - name: transactionType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ExpenseTransactionType"
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [newest, oldest, highest, lowest]
            default: newest
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A page of expenses matching the search, filters, and sort
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpensePage"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense with related records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRecord"
        "400":
          description: Invalid expense data or related record IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{expenseId}:
    parameters:
      - name: expenseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get an expense and its related records
      responses:
        "200":
          description: Expense detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRecord"
        "400":
          description: Invalid expense ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense and its label associations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated expense with related records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRecord"
        "400":
          description: Invalid expense data or related record IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      responses:
        "204":
          description: Expense deleted
        "400":
          description: Invalid expense ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Currency:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          minLength: 3
          maxLength: 3
        name:
          type: string
          minLength: 1
          maxLength: 80
        symbol:
          type: string
          minLength: 1
          maxLength: 8
        decimalPlaces:
          type: integer
          minimum: 0
          maximum: 4
        isActive:
          type: boolean
      required:
        - id
        - code
        - name
        - symbol
        - decimalPlaces
        - isActive
    CurrencyAmount:
      type: object
      properties:
        currency:
          $ref: "#/components/schemas/Currency"
        total:
          type: string
          pattern: "^[0-9]+(\\.[0-9]{1,2})?$"
        count:
          type: integer
          minimum: 0
        share:
          type: number
          minimum: 0
          maximum: 1
          description: Share of the matching same-currency total. Never compares across currencies.
      required:
        - currency
        - total
        - count
...
    ExpensePaymentMethod:
      type: string
      enum: [cash, credit_card, debit_card, bank_transfer, upi, other]
    ExpenseTransactionType:
      type: string
      description: "expense = real consumption. payment = settling a debt, excluded from spending totals."
      enum: [expense, payment]
    ExpenseProject:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
        status:
          type: string
          enum: [active, archived]
      required: [id, name, color, status]
    ExpenseLocation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        countryCode:
          type: [string, "null"]
          minLength: 2
          maxLength: 2
        status:
          type: string
          enum: [active, archived]
      required: [id, name, slug, countryCode, status]
    ExpenseCategory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        parentId:
          type: [string, "null"]
          format: uuid
          description: Null for a top-level category.
        icon:
          type: [string, "null"]
        color:
          type: string
        status:
          type: string
          enum: [active, archived]
      required: [id, name, slug, parentId, icon, color, status]
    ExpenseLabel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
        status:
          type: string
          enum: [active, archived]
      required: [id, name, color, status]
...
    ExpenseRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: string
          pattern: "^[0-9]{1,12}(\\.[0-9]{1,2})?$"
        date:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        projectId:
          type: [string, "null"]
          format: uuid
        categoryId:
          type: string
          format: uuid
        labelIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        description:
          type: [string, "null"]
        paymentMethod:
          anyOf:
            - $ref: "#/components/schemas/ExpensePaymentMethod"
            - type: "null"
        notes:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        project:
          anyOf:
            - $ref: "#/components/schemas/ExpenseProject"
            - type: "null"
        category:
          $ref: "#/components/schemas/ExpenseCategory"
        labels:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseLabel"
        currency:
          $ref: "#/components/schemas/Currency"
        location:
          $ref: "#/components/schemas/ExpenseLocation"
        transactionType:
          $ref: "#/components/schemas/ExpenseTransactionType"
      required:
        - id
        - amount
        - date
        - projectId
        - categoryId
        - labelIds
        - description
        - paymentMethod
        - notes
        - createdAt
        - updatedAt
        - project
        - category
        - labels
        - currency
        - location
        - transactionType
    ExpensePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseRecord"
        total:
          type: integer
          description: Total expenses matching the current search and filters
        totals:
          type: array
          description: Sum of matching expense transactions grouped by currency
          items:
            $ref: "#/components/schemas/CurrencyAmount"
        paymentTotals:
          type: array
          description: Sum of matching payment transactions grouped by currency
          items:
            $ref: "#/components/schemas/CurrencyAmount"
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
      required: [items, total, totals, paymentTotals, limit, offset, hasMore]
...
    ExpenseInput:
      type: object
      additionalProperties: false
      properties:
        amount:
          type: string
          pattern: "^[0-9]{1,12}(\\.[0-9]{1,2})?$"
          maxLength: 15
        date:
          type: string
          format: date
        locationId:
          type: string
          format: uuid
        projectId:
          type: [string, "null"]
          format: uuid
        categoryId:
          type: string
          format: uuid
        labelIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        description:
          type: [string, "null"]
          maxLength: 1000
        paymentMethod:
          anyOf:
            - $ref: "#/components/schemas/ExpensePaymentMethod"
            - type: "null"
        transactionType:
          anyOf:
            - $ref: "#/components/schemas/ExpenseTransactionType"
            - type: "null"
        notes:
          type: [string, "null"]
          maxLength: 4000
        currencyId:
          type: string
          format: uuid
      required: [amount, date, categoryId, currencyId, locationId]
    ExpenseUpdate:
      type: object
      additionalProperties: false
      properties:
        amount:
          type: string
          pattern: "^[0-9]{1,12}(\\.[0-9]{1,2})?$"
          maxLength: 15
        date:
          type: string
          format: date
        locationId:
          type: string
          format: uuid
        projectId:
          type: [string, "null"]
          format: uuid
        categoryId:
          type: string
          format: uuid
        labelIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        description:
          type: [string, "null"]
          maxLength: 1000
        paymentMethod:
          anyOf:
            - $ref: "#/components/schemas/ExpensePaymentMethod"
            - type: "null"
        transactionType:
          anyOf:
            - $ref: "#/components/schemas/ExpenseTransactionType"
            - type: "null"
        notes:
          type: [string, "null"]
          maxLength: 4000
        currencyId:
          type: string
          format: uuid
...