# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExpensePaymentMethod } from './expensePaymentMethod';
import type { ExpenseTransactionType } from './expenseTransactionType';

export interface ExpenseUpdate {
  /**
     * @maxLength 15
     * @pattern ^[0-9]{1,12}(\.[0-9]{1,2})?$
     */
  amount?: string;
  date?: Date;
  locationId?: string;
  /** @nullable */
  projectId?: string | null;
  categoryId?: string;
  labelIds?: string[];
  /**
     * @maxLength 1000
     * @nullable
     */
  description?: string | null;
  paymentMethod?: ExpensePaymentMethod | null;
  transactionType?: ExpenseTransactionType | null;
  /**
     * @maxLength 4000
     * @nullable
     */
  notes?: string | null;
  currencyId?: string;
}